requirements.txt
fastapi
python-dotenv
google-generativeai
essentia
uvicorn
gunicorn
python-multipart
requests
openai
numpy